DiabetesAI_v2/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
supabase==2.4.6
langchain==0.2.5
langchain-openai==0.1.14
pydantic[email]==2.7.3
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
numpy==1.26.4
joblib==1.4.2
xgboost==2.0.3
scikit-learn==1.5.0
tensorflow==2.16.1
